pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "simclaw"
version = "0.1.0"
description = "STAR-CCM+ 特化 自律CAEエージェント"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "chromadb>=0.4.0",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=2.2.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "pypdf>=3.0.0",
    "playwright>=1.40.0",
    "httpx>=0.24.0",
    "python-frontmatter>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
simclaw = "simclaw.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["simclaw*"]
